+++ src/test/java/ua/orlov/hw10/TestStudent.java
@@ -4,8 +4,6 @@ import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 public class TestStudent {
-
-
     @Test
     public void testStudent() {
         Student student = new Student();
@@ -20,17 +18,3 @@ public class TestStudent {
     }
 }
 
-class TestAspirant {
-    @Test
-    public void testAspirantMynum() {
-        Student student = new Aspirant();
-        student.avaregeMark = 4.5;
-        Assertions.assertEquals(180, student.getScholarship());
-    }
-
-    @Test
-    public void testAspirant() {
-        Student student = new Aspirant();
-        Assertions.assertEquals(200, student.getScholarship());
-    }
-}
